src/test/java/kr/co/spring_data_jpa_study/study/repository/MemberJpaRepositoryTest.java
package kr.co.spring_data_jpa_study.study.repository;

import kr.co.spring_data_jpa_study.study.entity.Member;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import static org.assertj.core.api.Assertions.*;

@SpringBootTest
class MemberJpaRepositoryTest {

    @Autowired
    MemberJpaRepository memberJpaRepository;

    @Test
    @DisplayName("회원 저장")
//    @Transactional
//    @Rollback(value = false)
    void save() {
        Member member = new Member("테스트 이름");
        Member savedMember = memberJpaRepository.save(member);

        Member findMember = memberJpaRepository.find(savedMember.getId());

        assertThat(findMember.getId()).isEqualTo(member.getId());
        assertThat(findMember.getUsername()).isEqualTo(member.getUsername());
        assertThat(findMember).isEqualTo(member);
    }

    @Test
    @DisplayName("회원 조회")
    void find() {
        memberJpaRepository.find(1L);
    }
}